Rework close handling by sending close to server when necessary
Use the same implementation as in JS version.
=> https://github.com/socketio/engine.io-client/tree/3.5.x/lib

Without this patch, the client does not send the close packet on PONG
timeout. Consequently, the server does not see immediatrly that the
client has gone.

It fixes the commit 1cfd42c7f71b665ea0615fd0974c30cf84c81fb4
  https://github.com/veo-labs/python-engineio-client/commit/1cfd42c7f71b665ea0615fd0974c30cf84c81fb4

I think that Fred's patch was made for a good reason (do not send close
when not needed) but the wrong way.

Signed-off-by: Gaylord CHARLES <gaylord.charles@veo-labs.com>